requirements.txt
# TFG: Detección de Phishing y Smishing con NLP y ML
# Python 3.10+

# Data manipulation
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# NLP
nltk>=3.8.0
spacy>=3.6.0

# Deep Learning (opcional - para modelos avanzados)
torch>=2.0.0

# Transformers (opcional - para modelos BERT)
transformers>=4.35.0
datasets>=2.14.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
wordcloud>=1.9.0

# Jupyter
jupyter>=1.0.0
notebook>=7.0.0

# Web Application
streamlit>=1.28.0

# Database
SQLAlchemy>=2.0.0

# Authentication
bcrypt>=4.0.0

# Explainability
shap>=0.43.0
lime>=0.2.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
pyyaml>=6.0.0
tqdm>=4.66.0
requests>=2.31.0
